Reservation (not finished) CRUD

+++ src/main/java/com/example/ReservationAppBackEnd/customServiceCategory/service/CustomServiceCategoryService.java
@@ -3,6 +3,9 @@ package com.example.ReservationAppBackEnd.customServiceCategory.service;
 import com.example.ReservationAppBackEnd.customServiceCategory.domain.CustomServiceCategory;
 import com.example.ReservationAppBackEnd.customServiceCategory.repository.CustomServiceCategoryRepository;
 import com.example.ReservationAppBackEnd.customServiceCategory.api.CustomServiceCategoryRequest;
+import com.example.ReservationAppBackEnd.user.domain.User;
+import com.example.ReservationAppBackEnd.user.service.UserService;
+import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -10,14 +13,12 @@ import java.util.List;
 import java.util.Optional;
 
 @Service
+@RequiredArgsConstructor
 public class CustomServiceCategoryService {
 
     private final CustomServiceCategoryRepository customServiceCategoryRepository;
+    private final UserService userService;
 
-    @Autowired
-    public CustomServiceCategoryService(CustomServiceCategoryRepository customServiceCategoryRepository) {
-        this.customServiceCategoryRepository = customServiceCategoryRepository;
-    }
 
     public List<CustomServiceCategory> getAllCategories() {
         return customServiceCategoryRepository.findAll();
@@ -28,29 +29,52 @@ public class CustomServiceCategoryService {
     }
 
     public CustomServiceCategory createCategory(CustomServiceCategoryRequest categoryRequest) {
-        CustomServiceCategory category = CustomServiceCategory.builder()
-                .name(categoryRequest.name())
-                .build();
-
-        return customServiceCategoryRepository.save(category);
+        User user = userService.getLoggedUser();
+        if(user.getRole().equals("ADMIN")) {
+            CustomServiceCategory category = CustomServiceCategory.builder()
+                    .name(categoryRequest.name())
+                    .build();
+            return customServiceCategoryRepository.save(category);
+        }
+        else {
+            throw new RuntimeException("xd");
+        }
     }
 
     public CustomServiceCategory updateCategory(Long id, CustomServiceCategoryRequest categoryRequest) {
         Optional<CustomServiceCategory> optionalCategory = customServiceCategoryRepository.findById(id);
 
         if (optionalCategory.isPresent()) {
+            User user = userService.getLoggedUser();
+            if(user.getRole().equals("ADMIN")) {
             CustomServiceCategory category = optionalCategory.get();
             category.setName(categoryRequest.name());
 
             return customServiceCategoryRepository.save(category);
         } else {
+                throw new RuntimeException("Category not found with id: " + id);
+            }
+        }
+            else {
             throw new RuntimeException("Category not found with id: " + id);
         }
     }
 
     public void deleteCategory(Long id) {
-        customServiceCategoryRepository.deleteById(id);
-    }
+        Optional<CustomServiceCategory> optionalCategory = customServiceCategoryRepository.findById(id);
 
-    // Dodaj inne metody związane z operacjami CRUD na kategoriach usług, jeśli są potrzebne
+        if (optionalCategory.isPresent()) {
+            User user = userService.getLoggedUser();
+            if (user.getRole().equals("ADMIN")) {
+                customServiceCategoryRepository.deleteById(id);
+            } else {
+                throw new RuntimeException("Category not found with id: " + id);
+            }
+        } else {
+            throw new RuntimeException("Category not found with id: " + id);
+
+        }
+
+    }
+            // Dodaj inne metody związane z operacjami CRUD na kategoriach usług, jeśli są potrzebne
 }
